moved signIn and signOut methods to be globally available, now working on cleaning up the code with an API object
