+++ client/src/App.js
@@ -1,6 +1,8 @@
 import React, {useState } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import Cookies from 'js-cookie';
+import btoa from 'btoa';
+import axios from 'axios';
 
 
 // Components
@@ -15,15 +17,25 @@ import UserSignOut from './components/UserSignOut';
 import PrivateRoute from './components/PrivateRoute';
 
 
+import Api from './Api';
+
 const App = () => {
 	const [user, setUser] = useState(Cookies.getJSON('authenticatedUser') || {} );
 	const [statePassword, setStatePassword] = useState('');
+
+	const api = new Api();
+	const signIn = api.signIn
+
+	const signOut = () => {
+		setUser({});
+		Cookies.remove('authenticatedUser');
+	}
 
 	return (
 		<div className="App">
 
 			<BrowserRouter>
-				<Header authenticatedUser={user} signOut={() => <UserSignOut setUser={setUser} />}/>
+				<Header authenticatedUser={user} signOut={signOut} />
 				<Switch>
 					{/* Protected Routes*/}
 					<PrivateRoute exact path="/courses/create" component={CreateCourse} authenticatedUser={user} statePassword={statePassword} />
@@ -33,7 +45,7 @@ const App = () => {
 					<Route exact path="/courses" component={Courses} />
 					<Route exact path="/signup" component={UserSignUp} />
 					<Route exact path="/courses/:id" render={()=> <CourseDetail authenticatedUser={user} statePassword={statePassword} />} />
-					<Route exact path="/signin" render={()=> <UserSignIn signIn={setUser} setStatePassword={setStatePassword} />} />
+					<Route exact path="/signin" render={()=> <UserSignIn signIn={signIn} setStatePassword={setStatePassword} />} />
 				</Switch>
 			</BrowserRouter>
 		</div>
@@ -41,3 +53,25 @@ const App = () => {
 }
 
 export default App;
+
+// const signIn = (userEmail, userPassword) => {
+    //     const encodedCredentials = btoa(`${userEmail}:${userPassword}`);
+    //     let config = {
+    //         method: 'get',
+    //         url: 'http://localhost:5000/api/users',
+    //         headers: { 
+    //           'Authorization': `Basic ${encodedCredentials}`
+    //         }
+    //       };
+          
+    //       axios(config)
+    //       .then(response => {
+    //         setUser(response.data);
+    //         Cookies.set('authenticatedUser', JSON.stringify(response.data), { expires: 1 });
+    //       })
+    //       .catch(function (error) {
+    //         console.log(error);
+    //       });
+    
+    //       setStatePassword(userPassword); // this adds the password to application state
+    // }
